Implement wikimedia/qunit coding style (#78)
Uses the eslint-plugin-qunit plugin by @platinumazure, and inherits
from the qunit/recommended and qunit/two rulesets.

In addition, we also follow these rules:
* qunit/no-assert-equal
* qunit/no-early-return
* qunit/no-negated-ok

The qunit/require-expect had to be turned off explicitly because the
default "recommended" set turns it on, however this is considered
a dated practice within Wikimedia. See also issue
https://github.com/platinumazure/eslint-plugin-qunit/issues/69.

Fixes #68.