Breaking: allow style lints during development
This patch the lowers all style concerns from errors to warnings. The
change allows users to focus on true development concerns-- programmer
errors and functionality instead of prettiness and whitespace
distractions prior to committing. ESLint is excellent at flagging common
JavaScript mishaps that are true bugs but when style concerns are
elevated to the same level as legitimate issues, the latter are much
more difficult to spot.

Since stylish readability is very important as soon as a commit is ready
to be reviewed and any time thereafter, upgrading clients are strongly
encouraged to specify `--max-warnings 0` in their pre-commit tests (or
use the supplied tool new to this patch, eslint-wikimedia).

Although this change in itself isn't breaking since it's lowering
severity, most clients are expected to want to change the way they
invoke ESLint to avoid unwanted style offenders creeping in.