Minor formatting change to Changelog (#149)
I copied this style from qunitjs originally, and thought it was
smart to use '##' instead of '=======' because it saved a line
and looked more "markdown"-ish (despite both being widely supported).

Now, having done QUnit releasing recently, I realise this was a
mistake because it means the very useful tool "git-changelog", which
QUnit use, doesn't work as expected.

See <https://github.com/tj/git-extras>, available on Homebrew.

Removing the first heading and using === makes it work, the rest
was already conforming (naturally). It does support any file
by the name of /(Change|History).*/i, so that's fine as well.