Tests: Separate "// Rule:" comments into Valid & Off
"// Valid:" comments are not asserted in any way. They just
show which rule we are giving an valid example of.

"// Off:" comments are when a rule has explicitly been turned
off and we will assert that it is present in the valid fixture.

We can now assert that enabled rules are never documented as "Off:".