Add config for TypeScript files
This takes over the config from https://github.com/wmde/eslint-config-wikimedia-typescript/
which was explicitly only every intended to be a provisional set of
configuration.

This temporary configuration led to problems when its version of this
configuration here had to be kept up-to-date or automatic upgrades with
LibraryUpgrader would fail, which happened regularly.

This adds the configuration here as a `.js` file, because that allows us
to retain the comments in the original file, which are still meaningful.
Precedence for `.js` config files already exists in the vue3/ directory.

@typescript-eslint/eslint-plugin and @typescript-eslint/parser are added
as direct dependencies in accordance with all the other dependencies
that are also direct dependencies like `eslint-plugin-qunit` and so one.

Bug: T301431